+++ tracing-subscriber/src/filter/mod.rs
@@ -8,15 +8,15 @@
 //! [`Layer`]: crate::layer
 #[cfg(feature = "env-filter")]
 mod env;
-#[cfg(feature = "registry")]
-mod layer_filters;
 mod level;
+#[cfg(feature = "registry")]
+mod subscriber_filters;
 
 #[cfg(not(feature = "registry"))]
 pub(crate) use self::has_plf_stubs::*;
 #[cfg(feature = "registry")]
 #[cfg_attr(docsrs, doc(cfg(feature = "registry")))]
-pub use self::layer_filters::*;
+pub use self::subscriber_filters::*;
 
 pub use self::level::{LevelFilter, ParseError as LevelParseError};
